+++ ProgettoOOPMeteo/src/main/java/it/univpm/ProgettoOOPMeteo/Operazioni/Statistiche.java
@@ -25,14 +25,14 @@ public class Statistiche {
 	  * @param fine
 	  * @return mediaVento
 	  */
-	 public Double mediaVento (Citta citta, int inizio, int fine){
+	 public float mediaVento (Citta citta, int inizio, int fine){
 		
 			float somma=0;
 		    
 			for (int i = inizio; i < fine ; i++) {
 		    	somma += citta.getPosVelocita(i);
 		    }
-			return (double) (somma/(fine - inizio + 1));
+			return (somma/(fine - inizio + 1));
 		}
 	
 	 /**
@@ -67,7 +67,7 @@ public class Statistiche {
 	
 	float somma=0; 
 	
-	Double media = mediaVento(citta, inizio, fine);
+	float media = mediaVento(citta, inizio, fine);
 	
 	for (int i=inizio; i<=fine; i++) {
 		somma += (citta.getPosVelocita(i) - media)*(citta.getPosVelocita(i) - media);
